Avoid panic caused by stdlib context package errors
