+++ transport/go16.go
@@ -35,7 +35,10 @@
 package transport
 
 import (
+	"fmt"
 	"net"
+
+	"google.golang.org/grpc/codes"
 
 	"golang.org/x/net/context"
 )
@@ -43,4 +46,15 @@ import (
 // dialContext connects to the address on the named network.
 func dialContext(ctx context.Context, network, address string) (net.Conn, error) {
 	return (&net.Dialer{Cancel: ctx.Done()}).Dial(network, address)
+}
+
+// ContextErr converts the error from context package into a StreamError.
+func ContextErr(err error) StreamError {
+	switch err {
+	case context.DeadlineExceeded:
+		return streamErrorf(codes.DeadlineExceeded, "%v", err)
+	case context.Canceled:
+		return streamErrorf(codes.Canceled, "%v", err)
+	}
+	panic(fmt.Sprintf("Unexpected error from context packet: %v", err))
 }
